# Conflicts:
#	frontend/src/components/adventureCreate/Step4Done.js
#	frontend/src/components/articleCreate/Step1Location.js
#	frontend/src/components/articleCreate/Step2Content.js

+++ backend/antenna/src/main/java/com/ssafy/antenna/domain/antenna/dto/DetailAntennaRes.java
@@ -1,4 +1,11 @@
 package com.ssafy.antenna.domain.antenna.dto;
 
-public record DetailAntennaRes(Long antennaId, int area, double lng, double lat, String w3w, String nearestPlace) {
+public record DetailAntennaRes(
+		Long antennaId,
+		Long area,
+		double lng,
+		double lat,
+		String w3w,
+		String nearestPlace
+) {
 }
